# public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM Paye API",
        "description": "API pour OM Paye - Système de paiement mobile",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/comptes/{num}/dashboard": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Obtenir le tableau de bord du compte",
                "operationId": "07b68a4d889a7f5b487359d35023c19d",
                "parameters": [
                    {
                        "name": "num",
                        "in": "path",
                        "description": "Numéro de téléphone du compte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "782345678"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informations du tableau de bord",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        },
                                        "compte": {
                                            "type": "object"
                                        },
                                        "transactions_recentes": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "statistiques": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comptes/{num}/solde": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Obtenir le solde du compte",
                "operationId": "36d755762c30c44e006e0e6027e8e174",
                "parameters": [
                    {
                        "name": "num",
                        "in": "path",
                        "description": "Numéro de téléphone du compte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "782345678"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solde du compte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "solde": {
                                            "description": "Solde en centimes",
                                            "type": "integer"
                                        },
                                        "solde_formate": {
                                            "description": "Solde formaté en FCFA",
                                            "type": "string"
                                        },
                                        "derniere_maj": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comptes/{num}/transactions": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Obtenir l'historique des transactions",
                "operationId": "1339f75ab14bc908db77927f9964f3c5",
                "parameters": [
                    {
                        "name": "num",
                        "in": "path",
                        "description": "Numéro de téléphone du compte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "782345678"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type de transaction",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "transfert",
                                "paiement",
                                "depot",
                                "retrait"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comptes/{num}/transfert": {
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Effectuer un transfert",
                "operationId": "a6444498aea55c690835c6b1c2ce0296",
                "parameters": [
                    {
                        "name": "num",
                        "in": "path",
                        "description": "Numéro de téléphone de l'expéditeur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "782345678"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone_destinataire",
                                    "montant",
                                    "password"
                                ],
                                "properties": {
                                    "telephone_destinataire": {
                                        "type": "string",
                                        "example": "783456789"
                                    },
                                    "montant": {
                                        "description": "Montant en centimes",
                                        "type": "integer",
                                        "example": "1000"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "motdepasse123"
                                    },
                                    "motif": {
                                        "type": "string",
                                        "example": "Transfert d'argent"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfert effectué",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "transaction": {
                                            "type": "object"
                                        },
                                        "solde_restant": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation"
                    },
                    "403": {
                        "description": "Mot de passe incorrect"
                    },
                    "404": {
                        "description": "Compte destinataire non trouvé"
                    },
                    "422": {
                        "description": "Solde insuffisant"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT pour l'authentification",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}